% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_anova.R
\name{av.aov}
\alias{av.aov}
\title{Convert an aov Object to Anytime-Valid aov (avaov)}
\usage{
\method{av}{aov}(model, g = 1, ...)
}
\arguments{
\item{model}{An \code{aov} object resulting from an ANOVA analysis.}

\item{g}{An integer precision parameter for anytime-valid inference. Default is 1.}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
An object of class \code{avaov} with anytime-valid p-values.
}
\description{
Converts an object of class \code{aov} to an anytime-valid version by setting the
precision parameter \code{g} as an attribute and updating the class.
}
