% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avsummary.R
\name{avsummary}
\alias{avsummary}
\title{An Anytime-Valid Summary for Linear Models}
\usage{
avsummary(lmfit, alpha = 0.05, phi = 1)
}
\arguments{
\item{lmfit}{an lm object resulting from running lm(y~.,data=)}

\item{alpha}{the desired 1-alpha confidence sequences}

\item{phi}{the precision of the mixture distribution on the standardized coefficients
(regression coefficients divided by the residual standard deviation).
If you are frequentist, a good rule of thumb is to set phi to 1/MDE^2 where
MDE is your estimate of the size of the standardized coefficients.
If you are Bayesian, set phi to be the precision of a Gaussian prior over
the standardized coefficients.}
}
\value{
an avsummary.lm object, very similar to a summary.lm object
}
\description{
avsummary is similar to summary with some important differences.
avsummary provides sequential p-values and confidence sequences which permit
analysis to be performed as frequently as desired with no need for multiple
testing corrections (coverage and Type-I guarantees are maintained under
continuous monitoring). For further details see
https://arxiv.org/abs/2210.08589
}
