% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_lm.R
\name{summary.avlm}
\alias{summary.avlm}
\title{Summary Method for Anytime-Valid lm (avlm) Objects}
\usage{
\method{summary}{avlm}(object, ...)
}
\arguments{
\item{object}{An \code{avlm} object generated by \code{av.lm}.}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
A summary object of class \code{summary.avlm} that includes updated p-values for the coefficients.
}
\description{
Computes a summary for an \code{avlm} object, a linear model enhanced with anytime-valid inference.
In addition to the standard \code{lm} summary statistics, the p-values for the coefficient tests are
recalculated using an anytime-valid method.
}
